+++ apps/docs/source.config.ts
@@ -8,7 +8,7 @@ import {
 import { transformerTwoslash } from 'fumadocs-twoslash';
 import { createFileSystemTypesCache } from 'fumadocs-twoslash/cache-fs';
 import remarkMath from 'remark-math';
-import { fileGenerator, remarkDocGen, remarkInstall } from 'fumadocs-docgen';
+import { remarkInstall } from 'fumadocs-docgen';
 import { remarkTypeScriptToJavaScript } from 'fumadocs-docgen/remark-ts2js';
 import rehypeKatex from 'rehype-katex';
 import { z } from 'zod';
@@ -17,6 +17,7 @@ import {
   remarkSteps,
 } from 'fumadocs-core/mdx-plugins';
 import { remarkAutoTypeTable } from 'fumadocs-typescript';
+import { transformerRemoveNotationEscape } from '@shikijs/transformers';
 
 export const docs = defineDocs({
   docs: {
@@ -63,31 +64,17 @@ export default defineConfig({
         transformerTwoslash({
           typesCache: createFileSystemTypesCache(),
         }),
-        {
-          name: 'transformers:remove-notation-escape',
-          code(hast) {
-            for (const line of hast.children) {
-              if (line.type !== 'element') continue;
-
-              const lastSpan = line.children.findLast(
-                (v) => v.type === 'element',
-              );
-
-              const head = lastSpan?.children[0];
-              if (head?.type !== 'text') continue;
-
-              head.value = head.value.replace(/\[\\!code/g, '[!code');
-            }
-          },
-        },
+        transformerRemoveNotationEscape(),
       ],
+    },
+    remarkCodeTabOptions: {
+      parseMdx: true,
     },
     remarkPlugins: [
       remarkSteps,
       remarkMath,
       remarkAutoTypeTable,
       [remarkInstall, { persist: { id: 'package-manager' } }],
-      [remarkDocGen, { generators: [fileGenerator()] }],
       remarkTypeScriptToJavaScript,
     ],
     rehypePlugins: (v) => [rehypeKatex, ...v],